requirements.txt
selenium >=4.27
python-decouple >= 3.8